refactor: move parseTheme to src/util/util.js
